requirements.txt
zoomus
python-telegram-bot
python-dotenv
